adds paynamics.html, paynamics.tsx; reverts back index.tsx and index.html
