+++ src/paynamics.html
@@ -4,11 +4,11 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>SleekyPay</title>
+    <title>Paynamics</title>
     <link rel="stylesheet" href="./index.scss">
 </head>
 <body>
     <div id="root"></div>
 </body>
-<script src="./index.tsx"></script>
+<script src="./paynamics.tsx"></script>
 </html>
